updating opinionated folder, adding a reconciler example
